+++ ops/overrides.nix
@@ -1,135 +1,11 @@
 {
 dev = {
-  flatparse = {
-  meta = {
-    sha256 = "0y6axksh2hqp8v58676a7zmwf0in7v6hmyfv8sfdx4x0acq2vjhr";
-    ver = "0.5.0.1";
-  };
-  drv = { mkDerivation, attoparsec, base, bytestring, containers, gauge
-, hspec, HUnit, integer-gmp, lib, megaparsec, parsec, primitive
-, QuickCheck, quickcheck-instances, template-haskell, utf8-string
-}:
-mkDerivation {
-  pname = "flatparse";
-  version = "0.5.0.1";
-  src = /nix/store/xkg9z7kvk4a3v2dfl0mh3sz50wl56srx-source;
-  libraryHaskellDepends = [
-    base bytestring containers integer-gmp template-haskell utf8-string
-  ];
-  testHaskellDepends = [
-    base bytestring hspec HUnit QuickCheck quickcheck-instances
-    utf8-string
-  ];
-  benchmarkHaskellDepends = [
-    attoparsec base bytestring gauge integer-gmp megaparsec parsec
-    primitive utf8-string
-  ];
-  homepage = "https://github.com/AndrasKovacs/flatparse#readme";
-  description = "High-performance parsing from strict bytestrings";
-  license = lib.licenses.mit;
-}
-;
-}
-;
 };
 ghc92 = {
-  flatparse = {
-  meta = {
-    sha256 = "0y6axksh2hqp8v58676a7zmwf0in7v6hmyfv8sfdx4x0acq2vjhr";
-    ver = "0.5.0.1";
-  };
-  drv = { mkDerivation, attoparsec, base, bytestring, containers, gauge
-, hspec, HUnit, integer-gmp, lib, megaparsec, parsec, primitive
-, QuickCheck, quickcheck-instances, template-haskell, utf8-string
-}:
-mkDerivation {
-  pname = "flatparse";
-  version = "0.5.0.1";
-  src = /nix/store/xkg9z7kvk4a3v2dfl0mh3sz50wl56srx-source;
-  libraryHaskellDepends = [
-    base bytestring containers integer-gmp template-haskell utf8-string
-  ];
-  testHaskellDepends = [
-    base bytestring hspec HUnit QuickCheck quickcheck-instances
-    utf8-string
-  ];
-  benchmarkHaskellDepends = [
-    attoparsec base bytestring gauge integer-gmp megaparsec parsec
-    primitive utf8-string
-  ];
-  homepage = "https://github.com/AndrasKovacs/flatparse#readme";
-  description = "High-performance parsing from strict bytestrings";
-  license = lib.licenses.mit;
-}
-;
-}
-;
 };
 ghc94 = {
-  flatparse = {
-  meta = {
-    sha256 = "0y6axksh2hqp8v58676a7zmwf0in7v6hmyfv8sfdx4x0acq2vjhr";
-    ver = "0.5.0.1";
-  };
-  drv = { mkDerivation, attoparsec, base, bytestring, containers, gauge
-, hspec, HUnit, integer-gmp, lib, megaparsec, parsec, primitive
-, QuickCheck, quickcheck-instances, template-haskell, utf8-string
-}:
-mkDerivation {
-  pname = "flatparse";
-  version = "0.5.0.1";
-  src = /nix/store/xkg9z7kvk4a3v2dfl0mh3sz50wl56srx-source;
-  libraryHaskellDepends = [
-    base bytestring containers integer-gmp template-haskell utf8-string
-  ];
-  testHaskellDepends = [
-    base bytestring hspec HUnit QuickCheck quickcheck-instances
-    utf8-string
-  ];
-  benchmarkHaskellDepends = [
-    attoparsec base bytestring gauge integer-gmp megaparsec parsec
-    primitive utf8-string
-  ];
-  homepage = "https://github.com/AndrasKovacs/flatparse#readme";
-  description = "High-performance parsing from strict bytestrings";
-  license = lib.licenses.mit;
-}
-;
-}
-;
 };
 ghc96 = {
-  flatparse = {
-  meta = {
-    sha256 = "0y6axksh2hqp8v58676a7zmwf0in7v6hmyfv8sfdx4x0acq2vjhr";
-    ver = "0.5.0.1";
-  };
-  drv = { mkDerivation, attoparsec, base, bytestring, containers, gauge
-, hspec, HUnit, integer-gmp, lib, megaparsec, parsec, primitive
-, QuickCheck, quickcheck-instances, template-haskell, utf8-string
-}:
-mkDerivation {
-  pname = "flatparse";
-  version = "0.5.0.1";
-  src = /nix/store/xkg9z7kvk4a3v2dfl0mh3sz50wl56srx-source;
-  libraryHaskellDepends = [
-    base bytestring containers integer-gmp template-haskell utf8-string
-  ];
-  testHaskellDepends = [
-    base bytestring hspec HUnit QuickCheck quickcheck-instances
-    utf8-string
-  ];
-  benchmarkHaskellDepends = [
-    attoparsec base bytestring gauge integer-gmp megaparsec parsec
-    primitive utf8-string
-  ];
-  homepage = "https://github.com/AndrasKovacs/flatparse#readme";
-  description = "High-performance parsing from strict bytestrings";
-  license = lib.licenses.mit;
-}
-;
-}
-;
   ghc-hs-meta = {
   meta = {
     sha256 = "1072k0a5svmw2f7sniq3z3n9ga4y8l5kx3qnsnbfsp4xhgcyq26i";
@@ -158,36 +34,5 @@ mkDerivation {
 hls = {
 };
 min = {
-  flatparse = {
-  meta = {
-    sha256 = "0y6axksh2hqp8v58676a7zmwf0in7v6hmyfv8sfdx4x0acq2vjhr";
-    ver = "0.5.0.1";
-  };
-  drv = { mkDerivation, attoparsec, base, bytestring, containers, gauge
-, hspec, HUnit, integer-gmp, lib, megaparsec, parsec, primitive
-, QuickCheck, quickcheck-instances, template-haskell, utf8-string
-}:
-mkDerivation {
-  pname = "flatparse";
-  version = "0.5.0.1";
-  src = /nix/store/xkg9z7kvk4a3v2dfl0mh3sz50wl56srx-source;
-  libraryHaskellDepends = [
-    base bytestring containers integer-gmp template-haskell utf8-string
-  ];
-  testHaskellDepends = [
-    base bytestring hspec HUnit QuickCheck quickcheck-instances
-    utf8-string
-  ];
-  benchmarkHaskellDepends = [
-    attoparsec base bytestring gauge integer-gmp megaparsec parsec
-    primitive utf8-string
-  ];
-  homepage = "https://github.com/AndrasKovacs/flatparse#readme";
-  description = "High-performance parsing from strict bytestrings";
-  license = lib.licenses.mit;
-}
-;
-}
-;
 };
 }